+++ specs/hd.h
@@ -399,3 +399,6 @@ ImU32 charIdToColor(CharacterIndex, float);
 CharacterIndex TextureIdToCharId(TextureId);
 void GenerateRoom(int32_t entrance_or_exit, LevelState *level_state,
                   int32_t x_start, int32_t y_start, int32_t room_idx);
+
+void mersenne_init_and_twist(uint32_t seed);
+uint32_t mersenne_random();